Splitting Custom Chat samples from the ChatGPT (Azure/OpenAI)
